registrar/requirements.txt
flatbuffers==2015.12.22.1
peewee==2.9.2
pyzmq==16.0.2
